feat: include looks in locations
